Update to latest version of tinygo
Since we moved to the vanilla gjson, the containerized builds have been
broken.
This commit changes the version of tinygo being used to be the latest
stable one. This is enough to fix the compilation errors.

Updating to tinygo 0.23 causes also a compilation error inside of the
container because the `git` executable is not found.
That's because upstream Go, since release 1.18, embeds the git
repository details inside of the final binaries, as reported inside of
the [release notes](https://tip.golang.org/doc/go1.18):

> The go command now embeds version control information in binaries.
> It includes the currently checked-out revision, commit time, and a flag
> indicating whether edited or untracked files are present. Version control
> information is embedded if the go command is invoked in a directory within a
> Git, Mercurial, Fossil, or Bazaar repository, and the main package and its
> containing main module are in the same repository.
> This information may be omitted using the flag -buildvcs=false.

We don't want to add the `git` binary to the official tinygo container
images, hence we're turning off this feature.
It's important to note that tinygo `build` command doesn't support the
`buildvcs` flag. However setting the flag via the `GOFLAGS` environment
variable works.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>