Make the code more idiomatic
Now that we learnt about easyjson and the Kubewarden Go SDK has been
updated to make use of it, we can revamp our demo policy to be more
idiomatic.

The policy now is making use of the Go Types our SDK provides, plus the
ones defined by the `github.com/kubewarden/k8s-objects`. Because of
that, the policy is no longer using `gjson`.

The unit tests have also been updated to significantly reduce the usage
of fixtures. I left one usage around, just in case someone wants to
continue using this approach.

Note: the policy can virtually target any kind of Kubernetes resource,
however I preferred to make it validate only Pod objects. As a matter
of fact, the `metadata.yml` was already targeting only Pod objects.

I also got rid of all the Ingress references because of that.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>